Create Comments
Once the data retrieval mechanism from Reddit is established, the subsequent step involves filtering for necessary data. For this case study, the focus is solely on collecting the top posts from the year 2023. An essential consideration is whether the score of each post is above 50, as this threshold helps identify significant posts. Posts with a score below 50 may not hold substantial importance.

Moreover, to ensure the relevance of the posts, it's essential to eliminate those with insufficient information. Hence, any post with a title containing fewer than 5 words will be excluded. This step helps remove posts that might be memes or jokes and ensures the inclusion of substantial content.

To enhance readability and comprehension, the Unix timestamps obtained through PRAW (Python Reddit API Wrapper) will be converted into human-readable time formats. The conversion will involve transforming the timestamps into a date format (year, month, and date) without specifying the exact time in hours, minutes, or seconds. This adjustment streamlines the representation of the temporal information for better understanding and analysis.